wip for global graph
